Massive cleanup pass

+++ src/main/java/downfall/events/shrines_evil/DuplicatorEvil.java
@@ -4,17 +4,16 @@ package downfall.events.shrines_evil;
 /*    */
 
 import com.megacrit.cardcrawl.cards.AbstractCard;
-/*    */ import com.megacrit.cardcrawl.core.CardCrawlGame;
+import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.core.Settings;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
-/*    */ import com.megacrit.cardcrawl.events.AbstractImageEvent;
-/*    */ import com.megacrit.cardcrawl.events.GenericEventDialog;
-/*    */ import com.megacrit.cardcrawl.helpers.CardLibrary;
+import com.megacrit.cardcrawl.events.AbstractImageEvent;
+import com.megacrit.cardcrawl.helpers.CardLibrary;
 import com.megacrit.cardcrawl.localization.EventStrings;
-/*    */ import com.megacrit.cardcrawl.screens.select.GridCardSelectScreen;
 import com.megacrit.cardcrawl.vfx.cardManip.ShowCardAndObtainEffect;
 import downfall.downfallMod;
-/*    */ import java.util.ArrayList;
+
+import java.util.ArrayList;
 
 /*    */
 /*    */ public class DuplicatorEvil extends AbstractImageEvent {
@@ -59,7 +58,7 @@ import downfall.downfallMod;
             AbstractDungeon.effectList.add(new com.megacrit.cardcrawl.vfx.cardManip.ShowCardAndObtainEffect(c, com.megacrit.cardcrawl.core.Settings.WIDTH * 0.25F, com.megacrit.cardcrawl.core.Settings.HEIGHT / 2.0F));
             cards.add(c.cardID);
 
-            if (AbstractDungeon.gridSelectScreen.selectedCards.size() > 1){
+            if (AbstractDungeon.gridSelectScreen.selectedCards.size() > 1) {
                 c = AbstractDungeon.gridSelectScreen.selectedCards.get(1).makeStatEquivalentCopy();
                 c.inBottleFlame = false;
                 c.inBottleLightning = false;
@@ -69,7 +68,7 @@ import downfall.downfallMod;
 
                 AbstractCard curse = CardLibrary.getCurse().makeStatEquivalentCopy();
                 cards.add(curse.cardID);
-                AbstractDungeon.effectList.add(new ShowCardAndObtainEffect(curse, (float) (Settings.WIDTH * .5F), (float) (Settings.HEIGHT / 2)));// 66
+                AbstractDungeon.effectList.add(new ShowCardAndObtainEffect(curse, Settings.WIDTH * .5F, (float) (Settings.HEIGHT / 2)));// 66
 
                 logMetricObtainCards(ID, "Desecrated", cards);
             } else {
